+++ client/src/hooks/Conversations/index.ts
@@ -5,8 +5,10 @@ export { default as useDefaultConvo } from './useDefaultConvo';
 export { default as useConversation } from './useConversation';
 export { default as useGenerateConvo } from './useGenerateConvo';
 export { default as useConversations } from './useConversations';
+export { default as useBookmarkSucess } from './useBookmarkSucess';
 export { default as useDebouncedInput } from './useDebouncedInput';
 export { default as useNavigateToConvo } from './useNavigateToConvo';
 export { default as useSetIndexOptions } from './useSetIndexOptions';
 export { default as useParameterEffects } from './useParameterEffects';
+export { default as useUpdateTagsInConvo } from './useUpdateTagsInConvo';
 export { default as useExportConversation } from './useExportConversation';
